Changes in various codes and added features to Recursos.cpp
## 06-02-2921

Changed multiple spelling mistakes in Los_ceros_no_valen.cpp. Changed the way I read a, b, c, as well as changed the place of the cycle that checks that a=>0. Changed the whole imaginary root part of the code to better accomodate the need of my teacher and express it properly.

Added the code Que_dias_aquellos.cpp which tells you which day of the week the first and last day of a month landed/will land in with the year and the month.

Added enumeration and cycles with examples in Recursos.cpp